+++ src/lambdaisland/hiccup.clj
@@ -8,6 +8,27 @@
             [garden.compiler :as gc]
             [clojure.string :as str]))
 
+(def kebab-case-tags
+  ;; from https://github.com/preactjs/preact-compat/issues/222
+  #{;; html
+    "accept-charset" "http-equiv"
+    ;; svg
+    "accent-height" "alignment-baseline" "arabic-form" "baseline-shift" "cap-height"
+    "clip-path" "clip-rule" "color-interpolation" "color-interpolation-filters"
+    "color-profile" "color-rendering" "fill-opacity" "fill-rule" "flood-color"
+    "flood-opacity" "font-family" "font-size" "font-size-adjust" "font-stretch"
+    "font-style" "font-variant" "font-weight" "glyph-name"
+    "glyph-orientation-horizontal" "glyph-orientation-vertical" "horiz-adv-x"
+    "horiz-origin-x" "marker-end" "marker-mid" "marker-start" "overline-position"
+    "overline-thickness" "panose-1" "paint-order" "stop-color" "stop-opacity"
+    "strikethrough-position" "strikethrough-thickness" "stroke-dasharray"
+    "stroke-dashoffset" "stroke-linecap" "stroke-linejoin" "stroke-miterlimit"
+    "stroke-opacity" "stroke-width" "text-anchor" "text-decoration" "text-rendering"
+    "underline-position" "underline-thickness" "unicode-bidi" "unicode-range"
+    "units-per-em" "v-alphabetic" "v-hanging" "v-ideographic" "v-mathematical"
+    "vert-adv-y" "vert-origin-x" "vert-origin-y" "word-spacing" "writing-mode"
+    "x-height"})
+
 (def block-level-tag?
   #{:head :body :meta :title :script :svg :iframe :style
     :link :address :article :aside :blockquote :details
@@ -16,6 +37,43 @@
 
 (defn- attr-map? [node-spec]
   (and (map? node-spec) (not (keyword? (:tag node-spec)))))
+
+(defn- kebab-in-html? [attr-str]
+  (or (contains? kebab-case-tags attr-str)
+      (str/starts-with? attr-str "data-")
+      (str/starts-with? attr-str "aria-")
+      (str/starts-with? attr-str "hx-")))
+
+(defn- kebab->camel [s]
+  (str/replace s #"-(\w)" (fn [[_ match]] (str/capitalize match))))
+
+(defn- camel->kebab [s]
+  (str/replace s #"([A-Z])" (fn [[_ match]] (str "-" (str/lower-case match)))))
+
+(defn convert-attribute-reagent-logic
+  [attr]
+  (cond
+    (string? attr)
+    attr
+    (keyword? attr)
+    (if (kebab-in-html? (name attr))
+      (name attr)
+      (kebab->camel (name attr)))))
+
+(defn convert-attribute-react-logic
+  [attr-str]
+  (let [kebab-str (camel->kebab attr-str)]
+    ;; not using kebab-in-html? here because
+    ;; React does not convert dataFoo to data-foo
+    ;; but does convert fontStretch to font-stretch
+    (if (kebab-case-tags kebab-str)
+      kebab-str
+      attr-str)))
+
+(defn convert-attribute [attr]
+  (->> attr
+       convert-attribute-reagent-logic
+       convert-attribute-react-logic))
 
 (defn- nodify [node-spec {:keys [newlines?] :as opts}]
   (cond
@@ -43,6 +101,12 @@
                              (into {} (filter val m))
                              {})
                     :content (enlive/flatmap #(nodify % opts) (if (attr-map? m) ms more))}
+              node (update node :attrs
+                           (fn [attrs]
+                             (->> attrs
+                                  (map (fn [[k v]]
+                                         [(convert-attribute k) v]))
+                                  (into {}))))
               node (if id (assoc-in node [:attrs :id] id) node)
               node (if (seq classes)
                      (update-in node
@@ -51,13 +115,13 @@
                                   (concat classes (if (string? kls) [kls] kls))))
                      node)]
           (cond-> node
-            (map? (get-in node [:attrs :style]))
-            (update-in [:attrs :style] (fn [style]
-                                         (-> (gc/compile-css [:& style])
-                                             (str/replace #"^\s*\{|\}\s*$" "")
-                                             str/trim)))
-            (sequential? (get-in node [:attrs :class]))
-            (update-in [:attrs :class] #(str/join " " %))
+            (map? (get-in node [:attrs "style"]))
+            (update-in [:attrs "style"] (fn [style]
+                                          (-> (gc/compile-css [:& style])
+                                              (str/replace #"^\s*\{|\}\s*$" "")
+                                              str/trim)))
+            (sequential? (get-in node [:attrs "class"]))
+            (update-in [:attrs "class"] #(str/join " " %))
             (and newlines? (block-level-tag? tag))
             (->> (list "\n"))))